@@ -11,14 +11,12 @@ import android.location.LocationManager;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.text.Layout;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
-import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
@@ -27,16 +25,18 @@ import androidx.annotation.Nullable;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
 
-import com.fivespecial.ploking.Maps.BinLocation;
-import com.fivespecial.ploking.Maps.Calculation;
-import com.fivespecial.ploking.Maps.DataAdapter;
 import com.fivespecial.ploking.R;
+import com.fivespecial.ploking.base.BaseFragment;
+import com.fivespecial.ploking.maps.BinLocation;
+import com.fivespecial.ploking.maps.Calculation;
+import com.fivespecial.ploking.maps.DataAdapter;
 import com.naver.maps.geometry.LatLng;
+import com.naver.maps.map.CameraPosition;
 import com.naver.maps.map.LocationTrackingMode;
 import com.naver.maps.map.MapFragment;
 import com.naver.maps.map.NaverMap;
+import com.naver.maps.map.NaverMapOptions;
 import com.naver.maps.map.NaverMapSdk;
 import com.naver.maps.map.UiSettings;
 import com.naver.maps.map.overlay.InfoWindow;
@@ -49,14 +49,15 @@ import java.util.List;
 
 import static android.content.Context.LOCATION_SERVICE;
 
-public class GPSFragment extends Fragment {
+public class GPSFragment extends BaseFragment {
 
-    //TAG
-    protected static final String TAG = "GPSFragment";
-
+    public static GPSFragment newInstance(){
+        return new GPSFragment();
+    }
 
     //custom-info-window
     private static class InfoWindowAdapter extends InfoWindow.DefaultTextAdapter {
+
         private InfoWindowAdapter(@NonNull Context context) {
             super(context);
         }
@@ -73,111 +74,107 @@ public class GPSFragment extends Fragment {
         }
     }
 
-    //garbage bins
-    Calculation calculation;
-    public List<BinLocation> binLocationList;
-    BinLocation binLocation;
-    TextView txtNearBin;
-    int binCount;
+    private static final String TAG = "GPSFragment";
+    private static final int LOCATION_PERMISSION_REQUEST_CODE = 1000;
 
+    private static final double DEFAULT_CAMERA_LATITUDE = 35.232286;
+    private static final double DEFAULT_CAMERA_LONGITUDE = 129.085166;
+    private static final double DEFAULT_BIN_LATITUDE = 35.232286;
+    private static final double DEFAULT_BIN_LONGITUDE = 129.085166;
 
-    //txtTimer
-    TextView txtDistance;
+    // Garbage Bins
+    private Calculation calculation;
+    private List<BinLocation> binLocationList;
+    private TextView txtNearBin;
+    private int binCount = 0;
+
+    // TextTimer
+    private TextView txtDistance;
     private TextView txtTime;
-    TextView txtKcal;
+    private TextView txtKcal;
 
-    double currentLon=0;
-    double currentLat=0;
-    double lastLon = 0;
-    double lastLat = 0;
-    float distance_sum = 0;
+    private double currentLon = 0;
+    private double currentLat = 0;
+    private double lastLon = 0;
+    private double lastLat = 0;
+    private float distance_sum = 0;
 
-    //timer
-    private ImageButton startBtn, stopBtn, pauseBtn;
+    // Timer
+    private ImageButton startButton, stopButton, pauseButton;
     private Boolean isRunning = false;
     private Boolean tButtonclicked = false;
-    Thread timeThread;
+    private Thread timeThread;
 
-
-    //kcal
-    int weight = 70;
-    private float coef = 0.001f; // 1초당 운동계수
-    float calorie = 0;
-
-
-
+    // Calorie
+    private float calorie = 0;
     private FusedLocationSource locationSource;
-    private static final int LOCATION_PERMISSION_REQUEST_CODE = 1000;
-    private MapFragment mapFragment;
 
-    public GPSFragment() {
+    private View runLayout;
+    private View informLayout;
+    private Animation pauseAppear;
+    private Animation stopAppear;
 
-    }
-
-    public static GPSFragment newInstance(){
-        return new GPSFragment();
+    @Override
+    public int getResourceId() {
+        return R.layout.fragment_gps;
     }
 
     @Override
-    public void onCreate(Bundle saveInstanceState) {
-        super.onCreate(saveInstanceState);
-
-        binCount = 0;
-
-        //API를 호출해 클라이언트 ID를 지정
-        NaverMapSdk.getInstance(getActivity()).setClient(
-                new NaverMapSdk.NaverCloudPlatformClient("0fwsvimx0a"));
-
-    }
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @NonNull ViewGroup container, @Nullable Bundle savedInstanceState) {
-
-        View view;
-        view = inflater.inflate(R.layout.fragment_gps, null);
-
-
+    public void initComponent(View view) {
 
         calculation = new Calculation();
+        locationSource = new FusedLocationSource(this, LOCATION_PERMISSION_REQUEST_CODE);
 
-        //define views...
-        startBtn = view.findViewById(R.id.btn_fragment_four_start);
-        stopBtn = view.findViewById(R.id.btn_fragment_four_stop);
-        pauseBtn = view.findViewById(R.id.btn_fragment_four_pause);
-        View runLayout = view.findViewById(R.id.layout_running);
-        View informLayout = view.findViewById(R.id.layout_inform);
+        startButton = view.findViewById(R.id.btn_fragment_four_start);
+        stopButton = view.findViewById(R.id.btn_fragment_four_stop);
+        pauseButton = view.findViewById(R.id.btn_fragment_four_pause);
+        runLayout = view.findViewById(R.id.layout_running);
+        informLayout = view.findViewById(R.id.layout_inform);
 
         //animation
-        Animation pauseAppear = AnimationUtils.loadAnimation(getActivity(), R.anim.pausebtn_appear);
-        Animation stopAppear = AnimationUtils.loadAnimation(getActivity(), R.anim.stopbtn_appear);
+        pauseAppear = AnimationUtils.loadAnimation(getActivity(), R.anim.pausebtn_appear);
+        stopAppear = AnimationUtils.loadAnimation(getActivity(), R.anim.stopbtn_appear);
 
         txtNearBin = view.findViewById(R.id.tv_near_bin_notice);
-        txtDistance = (TextView) view.findViewById(R.id.tv_fragment_four_distance);
-        txtTime = (TextView) view.findViewById(R.id.tv_fragment_four_time);
-        txtKcal = (TextView) view.findViewById(R.id.tv_fragment_four_calorie);
+        txtDistance = view.findViewById(R.id.tv_fragment_four_distance);
+        txtTime = view.findViewById(R.id.tv_fragment_four_time);
+        txtKcal = view.findViewById(R.id.tv_fragment_four_calorie);
+    }
 
-        //naverMapfragment start
-        FragmentManager fm = getFragmentManager();
-        mapFragment = (MapFragment)this
-                .getChildFragmentManager()
-                .findFragmentById(R.id.map);
-        if(mapFragment == null){
-            mapFragment = MapFragment.newInstance();
-            fm.beginTransaction().add(R.id.fragmentBorc, mapFragment).commit();
+    @Override
+    public void setupImplementation() {
+        try {
+
+            //API를 호출해 클라이언트 ID를 지정
+            if(getActivity() != null) {
+                NaverMapSdk.getInstance(getActivity()).setClient(
+                        new NaverMapSdk.NaverCloudPlatformClient("0fwsvimx0a"));
+            }
+
+            // 초기 위치 설정
+            NaverMapOptions options = new NaverMapOptions()
+                    .camera(new CameraPosition(new LatLng(35.232286, 129.085166), 14));
+
+            MapFragment mapFragment = MapFragment.newInstance(options);
+            getFragmentManager().beginTransaction().add(R.id.fragmentBorc, mapFragment).commit();
+
+            //비동기로 NaverMap 객체를 가져옴. NaverMap 객체가 준비되면 callback 의 onMapReady(NaverMap) 호출됨.
+            mapFragment.getMapAsync(this::onMapReady);
+
+        } catch (Exception e) {
+            e.printStackTrace();
+            Log.e(TAG, e + "에러 발생");
         }
 
-        locationSource = new FusedLocationSource(this, LOCATION_PERMISSION_REQUEST_CODE);
-        //test
-        mapFragment.getMapAsync(this::onMapReady);
+        // 시작버튼 클릭 리스너
+        startButton.setOnClickListener((View v) -> {
 
-        startBtn.setOnClickListener((View v) -> {
-            v.setVisibility(View.GONE);
+            v.setVisibility(View.GONE); // start 버튼 숨김
             runLayout.setVisibility(View.VISIBLE);
             informLayout.setVisibility(View.VISIBLE);
 
-            pauseBtn.startAnimation(pauseAppear);
-            stopBtn.startAnimation(stopAppear);
+            pauseButton.startAnimation(pauseAppear);
+            stopButton.startAnimation(stopAppear);
 
             isRunning = true;
             timeThread = new Thread(new timeThread());
@@ -185,17 +182,18 @@ public class GPSFragment extends Fragment {
             tButtonclicked = true;
         });
 
-        stopBtn.setOnClickListener((View v) ->{
-            startBtn.setVisibility(View.VISIBLE);
+        // 스탑버튼 클릭 리스너
+        stopButton.setOnClickListener((View v) ->{
+
+            startButton.setVisibility(View.VISIBLE); // 시작버튼을 다시 보이게 함.
             runLayout.setVisibility(View.GONE);
             informLayout.setVisibility(View.GONE);
 
             SharedPreferences sFile = getActivity().getSharedPreferences("sFile", Context.MODE_PRIVATE);
-
             SharedPreferences.Editor editor = sFile.edit();
 
-            Float updateDistance = sFile.getFloat("distance", 0);
-            Float updateKcal = sFile.getFloat("Kcal", 0);
+            float updateDistance = sFile.getFloat("distance", 0);
+            float updateKcal = sFile.getFloat("Kcal", 0);
 
             updateDistance += distance_sum;
             updateKcal += calorie;
@@ -203,111 +201,143 @@ public class GPSFragment extends Fragment {
             editor.putFloat("distance", updateDistance);
             editor.putFloat("Kcal", updateKcal);
 
-            editor.commit();
-            editor.commit();
+            editor.apply();
 
             isRunning = false;
             timeThread.interrupt();
-//                txtTime.setText("00:00");
             distance_sum = 0;
-//                txtDistance.setText("0.0 m");
         });
 
-        pauseBtn.setOnClickListener((View v) -> {
+        // 멈춤 버튼 클릭 리스너
+        pauseButton.setOnClickListener((View v) -> {
             isRunning = !isRunning;
+
             if(isRunning){
-                pauseBtn.setImageResource(R.drawable.pause_button);
+                pauseButton.setImageResource(R.drawable.pause_button);
             }else{
                 tButtonclicked = true;
-                pauseBtn.setImageResource(R.drawable.start_button);
+                pauseButton.setImageResource(R.drawable.start_button);
             }
         });
 
-        //GPS start...
-        int permissionCheck = ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION);   //test
+        // GPS 시작..
+        try {
 
-        if(permissionCheck == PackageManager.PERMISSION_DENIED){
-            if(ActivityCompat.shouldShowRequestPermissionRationale(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION)){}  //test
-            else{
-                ActivityCompat.requestPermissions(getActivity(),                                    //test
-                        new String[] {Manifest.permission.ACCESS_FINE_LOCATION},
-                        LOCATION_PERMISSION_REQUEST_CODE);
+            boolean permissionCheck = ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_DENIED;
+            boolean permissionFineLocation = !ActivityCompat.shouldShowRequestPermissionRationale(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION);
+
+            if(permissionCheck && permissionFineLocation){
+                ActivityCompat.requestPermissions(
+                        getActivity(), // activity
+                        new String[] {Manifest.permission.ACCESS_FINE_LOCATION}, // permissions
+                        LOCATION_PERMISSION_REQUEST_CODE // requestCode
+                );
+
             }
+
+            LocationManager lm = (LocationManager) getActivity().getSystemService(LOCATION_SERVICE);
+            Location loc = null;
+
+            if (lm != null) {
+                lm.requestLocationUpdates(LocationManager.GPS_PROVIDER, 0,0, setLocationListener());
+                loc = lm.getLastKnownLocation(LocationManager.GPS_PROVIDER);
+            }
+
+            if(loc == null){
+                txtDistance.setText("No GPS location found");
+            } else{
+                // 현재 위도와 경도를 설정함.
+                currentLon = loc.getLongitude();
+                currentLat = loc.getLatitude();
+            }
+            // 마지막으로 측정된 위도, 경도를 저장
+            lastLat = currentLat;
+            lastLon = currentLon;
+
+        } catch (NullPointerException e) {
+            e.printStackTrace();
         }
-
-        LocationManager lm =(LocationManager) getActivity().getSystemService(LOCATION_SERVICE);
-        lm.requestLocationUpdates(lm.GPS_PROVIDER, 0,0, Loclist);
-        Location loc = lm.getLastKnownLocation(lm.GPS_PROVIDER);
-
-        if(loc==null){
-            txtDistance.setText("No GPS location found");
-        }
-        else{
-            //set Current latitude and longitude
-            currentLon=loc.getLongitude();
-            currentLat=loc.getLatitude();
-        }
-        //Set the last latitude and longitude
-        lastLat=currentLat;
-        lastLon=currentLon;
-        //GPS end...
-
-
-        return view;
     }
 
     @Override
-    public void onRequestPermissionsResult(int requestCode,
-                                           @NonNull String[] permissions,  @NonNull int[] grantResults) {
-        if (locationSource.onRequestPermissionsResult(
-                requestCode, permissions, grantResults)) {
+    public void onCreate(Bundle saveInstanceState) {
+        super.onCreate(saveInstanceState);
+
+        initLoadDB(); // onMapReady 에서 DbHelper 에 접근했던 것을 onCreate 에서 하는것으로 변경
+
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,  @NonNull int[] grantResults) {
+
+        if (locationSource.onRequestPermissionsResult(requestCode, permissions, grantResults)) {
             return;
         }
-        super.onRequestPermissionsResult(
-                requestCode, permissions, grantResults);
+
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
     }
 
     // 맵을 생성할 준비가 되었을 때 가장 먼저 호출
-    public void onMapReady(@NonNull NaverMap naverMap) {
+    private void onMapReady(@NonNull NaverMap naverMap) {
 
         double longitude, latitude;
         double dLat, dLong;
         final InfoWindow binsInfo = new InfoWindow();
 
-
-
         final LocationManager lm = (LocationManager) getActivity().getSystemService(LOCATION_SERVICE);
         Location location;
 
-        //uisetting
+        try {
+            //position 속성을 지정하면 위치 오버레이의 좌표를 변경할 수 있다.
+            //처음 생성된 위치 오버레이는 카메라의 초기 좌표에 위치해 있다.
+            location = lm.getLastKnownLocation(LocationManager.GPS_PROVIDER);
+            latitude = location.getLatitude();
+            longitude = location.getLongitude();
+        }
+        catch(SecurityException | NullPointerException e){
+            Log.v(TAG, "CANNOT be find GPS signal.");
+            latitude = DEFAULT_CAMERA_LATITUDE;
+            longitude = DEFAULT_CAMERA_LONGITUDE;
+        }
+
+        Marker marker = new Marker();
+        marker.setPosition(new LatLng(latitude, longitude));
+
+        // uisetting
         UiSettings uiSettings = naverMap.getUiSettings();
+        uiSettings.setLocationButtonEnabled(true);
 
         naverMap.setLocationSource(locationSource);
-
-        LocationOverlay locationOverlay = naverMap.getLocationOverlay();
-        locationOverlay.setVisible(true);
-
-        //marker
-        initLoadDB();
-
         naverMap.setOnMapClickListener((coord, point) -> {
             binsInfo.close();
         });
+        naverMap.getLocationOverlay().setVisible(true);
+        naverMap.setLocationTrackingMode(LocationTrackingMode.Follow);
 
-        for(int i = 0; i < binLocationList.size(); i++){
-            binLocation = binLocationList.get(i);
+        // 오버레이 : 지리적 정보를 시작적으로 나타내느 요소. 개발자가 지도 위에 자유롭게 배치할 수 있음.
+        // 마커, 정보창(InfoWindow), 셰이프 등
+        LocationOverlay locationOverlay = naverMap.getLocationOverlay();
+        locationOverlay.setVisible(true);
+        locationOverlay.setBearing(90); // 방위를 지정함.
+        locationOverlay.setIconWidth(100); // 아이콘의 너비를 지정함
+        locationOverlay.setIconHeight(100); // 아이콘의 높이를 지정함.
+        locationOverlay.setPosition(new LatLng(latitude, longitude)); // 위치 지정
 
-            dLat = Double.parseDouble(binLocation.latitude);
-            dLong = Double.parseDouble(binLocation.longitude);
+        for(BinLocation binLocation : binLocationList){
 
-//            final InfoWindow binsInfo = new InfoWindow();
-//            binsInfo.setPosition(new LatLng(dLat, dLong));
+            if(binLocation.getLatitude() != null && binLocation.getLongitude() != null) {
+                dLat = binLocation.getLatitude();
+                dLong = binLocation.getLongitude();
+            } else {
+                dLat = DEFAULT_BIN_LATITUDE;
+                dLong = DEFAULT_BIN_LONGITUDE;
+            }
+
             binsInfo.setAdapter(new InfoWindowAdapter(getActivity()));
             binsInfo.setOnClickListener(overlay -> {
                 binsInfo.close();
                 return true;
             });
-
 
             Marker bins = new Marker();
             bins.setPosition(new LatLng(dLat, dLong));
@@ -320,127 +350,94 @@ public class GPSFragment extends Fragment {
             bins.setMap(naverMap);
         }
 
-
-
-        try {
-            location = lm.getLastKnownLocation(LocationManager.GPS_PROVIDER);
-            longitude = location.getLongitude();
-            latitude = location.getLatitude();
-            //position 속성을 지정하면 위치 오버레이의 좌표를 변경할 수 있다.
-            //처음 생성된 위치 오버레이는 카메라의 초기 좌표에 위치해 있다.
-        }
-        catch(SecurityException | NullPointerException e){
-            Log.v(TAG, "CANNOT be find GPS signal.");
-            longitude = 129.04378;
-            latitude = 35.237396;
-        }
-        binCount = calculation.NearBins(binLocationList, latitude, longitude);
+        binCount = calculation.nearBins(binLocationList, latitude, longitude);
         txtNearBin.setText(getString(R.string.near_bin_notice, binCount));
-        locationOverlay.setPosition(new LatLng(latitude, longitude));
-
-        Marker marker = new Marker();
-        marker.setPosition(new LatLng(latitude, longitude));
-
-
-        //bearing 속성을 지정하면 위치 오버레이의 베어링을 변경할 수 있다.
-        // flat이 true인 마커의 angle과 유사하게 아이콘이 지도를 기준으로 회전한다.
-        locationOverlay.setBearing(90);
-
-        locationOverlay.setIconWidth(100);
-        locationOverlay.setIconHeight(100);
-
-        naverMap.setLocationTrackingMode(LocationTrackingMode.Follow);
-
-        //uiSettings...
-        uiSettings.setLocationButtonEnabled(true);
 
     }
 
-    LocationListener Loclist = new LocationListener(){
 
-        @Override
-        public void onLocationChanged(Location location) {
+    private LocationListener setLocationListener() {
 
-            int binCount = 0;
+        return new LocationListener(){
 
-            try{
-                //start location manager
-                LocationManager lm =(LocationManager) getActivity().getSystemService(LOCATION_SERVICE);
+            @Override
+            public void onLocationChanged(Location location) {
 
-                //Request new location
-                lm.requestLocationUpdates(lm.GPS_PROVIDER, 0,0, Loclist);
+                try{
+                    //start location manager
+                    LocationManager lm =(LocationManager) getActivity().getSystemService(LOCATION_SERVICE);
 
-                //Get new location
-                Location loc = lm.getLastKnownLocation(lm.GPS_PROVIDER);
+                    //Request new location
+                    lm.requestLocationUpdates(lm.GPS_PROVIDER, 0,0, this);
 
-                //get the current lat and long
-                currentLat = loc.getLatitude();
-                currentLon = loc.getLongitude();
+                    //Get new location
+                    Location loc = lm.getLastKnownLocation(lm.GPS_PROVIDER);
 
-                binCount = calculation.NearBins(binLocationList, currentLat, currentLon);
+                    //get the current lat and long
+                    currentLat = loc.getLatitude();
+                    currentLon = loc.getLongitude();
 
-                txtNearBin.setText(getString(R.string.near_bin_notice, binCount));
+                    binCount = calculation.nearBins(binLocationList, currentLat, currentLon);
+
+                    txtNearBin.setText(getString(R.string.near_bin_notice, binCount));
 
 
-                if(isRunning){
-                    if(tButtonclicked){
+                    if(isRunning){
+                        if(tButtonclicked){
+                            lastLat = currentLat;
+                            lastLon = currentLon;
+                            tButtonclicked = false;
+                        }
+
+                        Location locationA = new Location("point A");
+                        locationA.setLatitude(lastLat);
+                        locationA.setLongitude(lastLon);
+
+                        Location locationB = new Location("point B");
+                        locationB.setLatitude(currentLat);
+                        locationB.setLongitude(currentLon);
+
+                        float distanceMeters = locationA.distanceTo(locationB);
+                        Log.v(TAG, String.format("%f", lastLon));
+                        Log.v(TAG, String.format("%f", currentLon));
+
+                        distance_sum += distanceMeters;
+
                         lastLat = currentLat;
                         lastLon = currentLon;
-                        tButtonclicked = false;
+
+                        txtDistance.setText(String.format("%.1f m",distance_sum ));
                     }
-
-                    Location locationA = new Location("point A");
-                    locationA.setLatitude(lastLat);
-                    locationA.setLongitude(lastLon);
-
-                    Location locationB = new Location("point B");
-                    locationB.setLatitude(currentLat);
-                    locationB.setLongitude(currentLon);
-
-                    float distanceMeters = locationA.distanceTo(locationB);
-                    Log.v(TAG, String.format("%f", lastLon));
-                    Log.v(TAG, String.format("%f", currentLon));
-
-                    distance_sum += distanceMeters;
-
-                    lastLat = currentLat;
-                    lastLon = currentLon;
-
-                    txtDistance.setText(String.format("%.1f m",distance_sum ));
+                }
+                catch(SecurityException | NullPointerException e){
+                    e.printStackTrace();
                 }
             }
-            catch(SecurityException | NullPointerException e){
-            }
-        }
 
-        @Override
-        public void onProviderDisabled(String provider) {
-        }
+            @Override
+            public void onProviderDisabled(String provider) { }
 
+            @Override
+            public void onProviderEnabled(String provider) { }
 
+            @Override
+            public void onStatusChanged(String provider, int status, Bundle extras) { }
 
-        @Override
-        public void onProviderEnabled(String provider) {
-        }
-
-
-
-        @Override
-        public void onStatusChanged(String provider, int status, Bundle extras) {
-        }
-
-    };
+        };
+    }
 
     @SuppressLint("HandlerLeak")
-    Handler handler = new Handler() {
+    private Handler handler = new Handler() {
         @Override
         public void handleMessage(Message msg) {
+
+            //1000이 1초 1000*60 은 1분 1000*60*10은 10분 1000*60*60은 한시간
             int sec_sum = msg.arg1;
             int sec = msg.arg1 % 60;
             int min = msg.arg1 / 60;
-            //1000이 1초 1000*60 은 1분 1000*60*10은 10분 1000*60*60은 한시간
 
-            //소모 칼로리
+            int weight = 70; //소모 칼로리(kcal)
+            float coef = 0.001f; // 1초당 운동계수
             calorie = weight * coef * sec_sum;
 
             @SuppressLint("DefaultLocale") String result = String.format("%02d:%02d", min, sec);
@@ -457,7 +454,9 @@ public class GPSFragment extends Fragment {
             int i = 0;
 
             while (true) {
-                while (isRunning) { //일시정지를 누르면 멈춤
+
+                //일시정지를 누르면 멈춤
+                while (isRunning) {
                     Log.d("stopwatch", "while start");
                     Message msg = new Message();
                     msg.arg1 = i++;
@@ -468,12 +467,11 @@ public class GPSFragment extends Fragment {
                         Log.d("stopwatch", "sleep");
                     } catch (InterruptedException e) {
                         e.printStackTrace();
-                        getActivity().runOnUiThread(new Runnable(){     //test
-                            @Override
-                            public void run() {
-                                txtTime.setText("");
-                                txtTime.setText("00:00");
-                            }
+                        //test
+                        getActivity().runOnUiThread(() -> {
+
+                            txtTime.setText("");
+                            txtTime.setText("00:00");
                         });
                         return;
                     }
